+++ apps/common/src/main/scala/com/daml/network/util/CNNodeUtil.scala
@@ -196,7 +196,8 @@ object CNNodeUtil {
 
   // TODO(#6032): determine the best defaults here
   val defaultExtraTrafficPrice = BigDecimal(1.0) // extraTrafficPrice (in $/MB)
-  val defaultMinTopupAmount = 1_000L // minTopupAmount = 1KB
+  val defaultMinTopupAmount =
+    1_000_000L // minTopupAmount chosen such that at least 1$ is spent per top-up (too low and the output fees for its corresponding transfer can't even be paid)
 
   // These are dummy values only made use of by some unit tests.
   // The traffic control parameters are provided in the founding SV App config with the defaults in TrafficControlConfig